Chat room component with colors
